🌱 [scanner] refactor: split navigation/settings components (Part of #21375)
- Extract SidebarShell.tsx (988→492 LOC) into layout/sidebar/ submodules:
  - useSidebarPin.ts: pin/auto-hide logic
  - useSidebarResize.ts: resize handle logic
  - useSidebarDragDrop.ts: drag-and-drop handlers
  - SidebarNavItemRow.tsx: nav item rendering with badge/rename/drag support
  - SidebarClusterStatus.tsx: cluster health summary widget
  - SidebarActiveUsersFooter.tsx: viewer count + commit hash footer
  - SidebarCollapseControls.tsx: collapse/pin floating buttons

- Extract UpdateSettingsForm.tsx (916→493 LOC) into settings/ helpers:
  - UpdatePrereqRow.tsx: prerequisite status row
  - UpdateCommitList.tsx: recent commits expandable list
  - UpdateProgressBanners.tsx: update progress/cancelled/done/failed banners
  - UpdateVersionInfo.tsx: version info card
  - UpdateHowToSection.tsx: install instructions for dev/stable channels
  - UpdateReleaseNotes.tsx: collapsible release notes

All public exports and props APIs remain unchanged.
