+++ web/src/components/layout/SidebarShell.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines -- TODO: split this file (tracked by #15790) */
 /**
  * SidebarShell — Reusable sidebar infrastructure component.
  *
@@ -9,22 +8,15 @@
  *   - Enterprise compliance sidebar (EnterpriseSidebar.tsx)
  *   - Future white-label / partner portals
  */
-import { useState, useRef, useEffect, useCallback, Fragment } from 'react'
+import { useState, useRef, useEffect, Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
-import { NavLink, useNavigate, useLocation } from 'react-router-dom'
-import {
-  Plus, ChevronLeft, ChevronRight, CheckCircle2, AlertTriangle,
-  WifiOff, GripVertical, X, User, Pin, PinOff, Satellite, Loader2,
-  ChevronDown,
-} from 'lucide-react'
-import { iconRegistry } from '../../lib/icons'
+import { useNavigate, useLocation } from 'react-router-dom'
+import { Plus } from 'lucide-react'
 import { cn } from '../../lib/cn'
 import { ROUTES } from '../../config/routes'
-import { Tooltip } from '../ui/Tooltip'
 import { SnoozedCards } from './SnoozedCards'
 import {
   useSidebarConfig,
-  PROTECTED_SIDEBAR_IDS,
   SIDEBAR_COLLAPSED_WIDTH_PX,
   SIDEBAR_DEFAULT_WIDTH_PX,
 } from '../../hooks/useSidebarConfig'
@@ -37,17 +29,19 @@ import type { SnoozedRecommendation } from '../../hooks/useSnoozedRecommendation
 import type { SnoozedMission } from '../../hooks/useSnoozedMissions'
 import { useActiveUsers } from '../../hooks/useActiveUsers'
 import { useMissions } from '../../hooks/useMissions'
-import { DASHBOARD_CONFIGS } from '../../config/dashboards/index'
-import { emitSidebarNavigated, emitDashboardRenamed } from '../../lib/analytics'
-import { prefetchDashboard } from '../../lib/prefetchDashboard'
+import { emitDashboardRenamed } from '../../lib/analytics'
 import { useVersionCheck } from '../../hooks/useVersionCheck'
 import { useUpgradeState } from '../../hooks/useUpgradeState'
-import { STORAGE_KEY_GROUND_CONTROL_DASHBOARDS } from '../../lib/constants/storage'
-import { NAVBAR_HEIGHT_PX, SIDEBAR_CONTROLS_LEFT_OFFSET_PX } from '../../lib/constants/ui'
-import { safeGetJSON } from '../../lib/utils/localStorage'
-import { getSidebarCardCount } from './sidebarCardCount'
+import { NAVBAR_HEIGHT_PX } from '../../lib/constants/ui'
 import { moveFocusByKey } from '../../lib/a11y/rovingFocus'
 import { useEscapeLayer, useModalFocusTrap } from '../../lib/modals'
+import { useSidebarPin } from './sidebar/useSidebarPin'
+import { useSidebarResize } from './sidebar/useSidebarResize'
+import { useSidebarDragDrop } from './sidebar/useSidebarDragDrop'
+import { SidebarNavItemRow } from './sidebar/SidebarNavItemRow'
+import { SidebarClusterStatus } from './sidebar/SidebarClusterStatus'
+import { SidebarActiveUsersFooter } from './sidebar/SidebarActiveUsersFooter'
+import { SidebarCollapseControls } from './sidebar/SidebarCollapseControls'
 
 // ---------------------------------------------------------------------------
 // Types
@@ -140,36 +134,6 @@ const SIDEBAR_RESIZE_HANDLE_WIDTH_PX = 6
 /** Index of the primary (dashboard list) section — "Add more..." button renders after it */
 const PRIMARY_SECTION_INDEX = 0
 
-/** Map sidebar item href to dashboard config ID for card count display. */
-const HREF_TO_DASHBOARD_ID: Record<string, string> = {
-  [ROUTES.HOME]: 'main', [ROUTES.COMPUTE]: 'compute', [ROUTES.SECURITY]: 'security',
-  [ROUTES.GITOPS]: 'gitops', [ROUTES.STORAGE]: 'storage', [ROUTES.NETWORK]: 'network',
-  [ROUTES.EVENTS]: 'events', [ROUTES.ALERTS]: 'alerts', [ROUTES.WORKLOADS]: 'workloads', [ROUTES.OPERATORS]: 'operators',
-  [ROUTES.CLUSTERS]: 'clusters', [ROUTES.COMPLIANCE]: 'compliance', [ROUTES.COST]: 'cost',
-  [ROUTES.GPU_RESERVATIONS]: 'gpu', [ROUTES.NODES]: 'nodes', [ROUTES.DEPLOYMENTS]: 'deployments',
-  [ROUTES.PODS]: 'pods', [ROUTES.SERVICES]: 'services', [ROUTES.HELM]: 'helm',
-  [ROUTES.AI_ML]: 'ai-ml', [ROUTES.CI_CD]: 'ci-cd',
-  [ROUTES.LOGS]: 'logs', [ROUTES.DATA_COMPLIANCE]: 'data-compliance', [ROUTES.ARCADE]: 'arcade',
-  [ROUTES.DEPLOY]: 'deploy', [ROUTES.AI_AGENTS]: 'ai-agents',
-  [ROUTES.LLM_D_BENCHMARKS]: 'llm-d-benchmarks', [ROUTES.CLUSTER_ADMIN]: 'cluster-admin',
-  [ROUTES.INSIGHTS]: 'insights', [ROUTES.DRASI]: 'drasi',
-  [ROUTES.MULTI_TENANCY]: 'multi-tenancy', [ROUTES.ACMM]: 'acmm',
-}
-
-const CUSTOM_DASHBOARD_PREFIX = ROUTES.CUSTOM_DASHBOARD.replace(':id', '')
-
-function isGroundControlItem(href: string): boolean {
-  if (!href.startsWith(CUSTOM_DASHBOARD_PREFIX)) return false
-  const dashboardId = href.slice(CUSTOM_DASHBOARD_PREFIX.length)
-  const gcMapping = safeGetJSON<Record<string, unknown>>(STORAGE_KEY_GROUND_CONTROL_DASHBOARDS) ?? {}
-  return dashboardId in gcMapping
-}
-
-const SIDEBAR_AUTO_HIDE_MS = 2000
-const COLLAPSED_BADGE_MAX_COUNT = 99
-const COLLAPSED_BADGE_BASE_CLASS = 'absolute -top-1 -right-1 z-10 flex h-4 min-w-4 items-center justify-center rounded-full border border-background px-0.5 text-xs font-bold leading-none shadow-sm'
-const COLLAPSED_BADGE_DEFAULT_COLOR_CLASS = 'bg-primary text-primary-foreground'
-
 // ---------------------------------------------------------------------------
 // Component
 // ---------------------------------------------------------------------------
@@ -197,7 +161,6 @@ export function SidebarShell({
   const { t } = useTranslation()
   const navigate = useNavigate()
   const location = useLocation()
-  const autoHideTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
   const sidebarRef = useRef<HTMLElement | null>(null)
   const isMobileSidebarOpen = isMobile && config.isMobileOpen
   const isTopEscapeLayer = useEscapeLayer(isMobileSidebarOpen)
@@ -206,159 +169,58 @@ export function SidebarShell({
 
   // Close mobile sidebar on route change
   useEffect(() => {
-    if (isMobile) {
-      closeMobileSidebar()
-    }
+    if (isMobile) closeMobileSidebar()
   }, [location.pathname, isMobile, closeMobileSidebar])
 
   useEffect(() => {
-    if (isMobileSidebarOpen && sidebarRef.current) {
-      sidebarRef.current.scrollTop = 0
-    }
+    if (isMobileSidebarOpen && sidebarRef.current) sidebarRef.current.scrollTop = 0
   }, [isMobileSidebarOpen])
 
   useModalFocusTrap(sidebarRef, isMobileSidebarOpen)
 
   useEffect(() => {
     if (!isMobileSidebarOpen) return
-
     const handleKeyDown = (event: KeyboardEvent) => {
       if (event.key !== 'Escape' || !isTopEscapeLayer() || editingItemId !== null) return
       event.preventDefault()
       event.stopPropagation()
       closeMobileSidebar()
     }
-
     window.addEventListener('keydown', handleKeyDown)
     return () => window.removeEventListener('keydown', handleKeyDown)
   }, [closeMobileSidebar, editingItemId, isMobileSidebarOpen, isTopEscapeLayer])
 
-  // ---- Auto-hide: collapse sidebar when mouse leaves, expand on hover ----
-  const [isPinned, setIsPinned] = useState(() => {
-    try { return localStorage.getItem('sidebar-left-pinned') !== 'false' } catch { return true }
+  const isCollapsed = !isMobile && config.collapsed
+  const sidebarWidth = isCollapsed ? SIDEBAR_COLLAPSED_WIDTH_PX : (widthOverride ?? config.width ?? SIDEBAR_DEFAULT_WIDTH_PX)
+
+  const { isPinned, handleSidebarMouseEnter, handleSidebarMouseLeave, toggleSidebarPin } = useSidebarPin({
+    isCollapsed, isMobile, collapsed: config.collapsed, setCollapsed,
   })
 
-  const clearAutoHideTimer = useCallback(() => {
-    if (autoHideTimerRef.current) {
-      clearTimeout(autoHideTimerRef.current)
-      autoHideTimerRef.current = null
-    }
-  }, [])
+  const { isResizing, handleResizeStart, handleResizeKeyDown } = useSidebarResize({
+    sidebarWidth, widthOverride, configWidth: config.width, setWidth,
+    SIDEBAR_DEFAULT_WIDTH_PX, SIDEBAR_MIN_WIDTH_PX, SIDEBAR_MAX_WIDTH_PX, SIDEBAR_RESIZE_STEP_PX,
+  })
 
-  const handleSidebarMouseEnter = () => {
-    clearAutoHideTimer()
-    if (!isPinned && config.collapsed && !isMobile) {
-      setCollapsed(false)
-    }
-  }
+  const {
+    draggedItem, dragOverItem, dragSection,
+    handleDragStart, handleDragEnd, handleDragEnter,
+    handleDragLeave, handleDragOver, handleDrop,
+  } = useSidebarDragDrop({ primaryNav: config.primaryNav, secondaryNav: config.secondaryNav, reorderItems })
 
-  const handleSidebarMouseLeave = () => {
-    if (!isPinned && !isMobile) {
-      clearAutoHideTimer()
-      autoHideTimerRef.current = setTimeout(() => {
-        setCollapsed(true)
-      }, SIDEBAR_AUTO_HIDE_MS)
-    }
-  }
-
-  const toggleSidebarPin = () => {
-    setIsPinned(prev => {
-      const next = !prev
-      try { localStorage.setItem('sidebar-left-pinned', String(next)) } catch { /* ignore */ }
-      if (next) {
-        clearAutoHideTimer()
-        if (config.collapsed) {
-          setCollapsed(false)
-        }
-      } else if (!config.collapsed) {
-        autoHideTimerRef.current = setTimeout(() => setCollapsed(true), SIDEBAR_AUTO_HIDE_MS)
-      }
-      return next
-    })
-  }
-
-  const handleMobileBackdropClose = () => {
-    if (editingItemId !== null) return
-    closeMobileSidebar()
-  }
-
-  useEffect(() => () => clearAutoHideTimer(), [clearAutoHideTimer])
-
-  const isCollapsed = !isMobile && config.collapsed
-  const sidebarWidth = isCollapsed
-    ? SIDEBAR_COLLAPSED_WIDTH_PX
-    : (widthOverride ?? config.width ?? SIDEBAR_DEFAULT_WIDTH_PX)
-
-  // ---- Resize handle ----
-  const [isResizing, setIsResizing] = useState(false)
-  const resizeCleanupRef = useRef<(() => void) | null>(null)
-
-  const clampSidebarWidth = useCallback((nextWidth: number) => {
-    return Math.min(SIDEBAR_MAX_WIDTH_PX, Math.max(SIDEBAR_MIN_WIDTH_PX, nextWidth))
-  }, [])
-
-  // Clean up resize listeners on unmount to prevent leaks if mouseup never fires
-  useEffect(() => () => { resizeCleanupRef.current?.() }, [])
-
-  const handleResizeStart = (e: React.MouseEvent) => {
-    e.preventDefault()
-    setIsResizing(true)
-    const startX = e.clientX
-    const startWidth = widthOverride ?? config.width ?? SIDEBAR_DEFAULT_WIDTH_PX
-
-    const handleMouseMove = (moveEvent: MouseEvent) => {
-      const newWidth = clampSidebarWidth(startWidth + (moveEvent.clientX - startX))
-      setWidth(newWidth)
-    }
-
-    const handleMouseUp = () => {
-      setIsResizing(false)
-      document.removeEventListener('mousemove', handleMouseMove)
-      document.removeEventListener('mouseup', handleMouseUp)
-      document.body.style.cursor = ''
-      document.body.style.userSelect = ''
-      resizeCleanupRef.current = null
-    }
-
-    document.body.style.cursor = 'col-resize'
-    document.body.style.userSelect = 'none'
-    document.addEventListener('mousemove', handleMouseMove)
-    document.addEventListener('mouseup', handleMouseUp)
-    resizeCleanupRef.current = handleMouseUp
-  }
-
-  const handleResizeKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    let delta = 0
-
-    if (event.key === 'ArrowLeft') {
-      delta = -SIDEBAR_RESIZE_STEP_PX
-    } else if (event.key === 'ArrowRight') {
-      delta = SIDEBAR_RESIZE_STEP_PX
-    }
-
-    if (delta === 0) {
-      return
-    }
-
-    event.preventDefault()
-    setWidth(clampSidebarWidth(sidebarWidth + delta))
-  }
-
-  // ---- Drag and drop state ----
-  const [draggedItem, setDraggedItem] = useState<string | null>(null)
-  const [dragOverItem, setDragOverItem] = useState<string | null>(null)
-  const [dragSection, setDragSection] = useState<string | null>(null)
-  const dragCounter = useRef(0)
-
-  // ---- Cluster status counts ----
   const unreachableClusters = deduplicatedClusters.filter((c) => isClusterUnreachable(c)).length
   const healthyClusters = deduplicatedClusters.filter((c) => !isClusterUnreachable(c) && isClusterHealthy(c)).length
   const unhealthyClusters = deduplicatedClusters.length - healthyClusters - unreachableClusters
 
   // ---- Snoozed / swap handlers ----
-  const handleApplySwap = (_swap: SnoozedSwap) => { navigate(ROUTES.HOME) }
-  const handleApplyRecommendation = (_rec: SnoozedRecommendation) => { navigate(ROUTES.HOME) }
-  const handleApplyMission = (_mission: SnoozedMission) => { navigate(ROUTES.HOME) }
+  const handleApplySwap = (_s: SnoozedSwap) => navigate(ROUTES.HOME)
+  const handleApplyRecommendation = (_r: SnoozedRecommendation) => navigate(ROUTES.HOME)
+  const handleApplyMission = (_m: SnoozedMission) => navigate(ROUTES.HOME)
+
+  const handleMobileBackdropClose = () => {
+    if (editingItemId !== null) return
+    closeMobileSidebar()
+  }
 
   // ---- Inline rename handlers ----
   const handleDoubleClick = (item: SidebarNavItem, e: React.MouseEvent) => {
@@ -379,282 +241,9 @@ export function SidebarShell({
     setEditingName('')
   }
 
-  const handleClusterStatusClick = (status: 'healthy' | 'unhealthy' | 'unreachable') => {
-    navigate(`${ROUTES.CLUSTERS}?status=${status}`)
-  }
-
-  // ---- Drag handlers ----
-  const handleDragStart = (e: React.DragEvent, itemId: string, sectionId: string) => {
-    setDraggedItem(itemId)
-    setDragSection(sectionId)
-    e.dataTransfer.effectAllowed = 'move'
-    e.dataTransfer.setData('text/plain', itemId)
-    requestAnimationFrame(() => {
-      const target = e.target as HTMLElement
-      target.style.opacity = '0.5'
-    })
-  }
-
-  const handleDragEnd = (e: React.DragEvent) => {
-    const target = e.target as HTMLElement
-    target.style.opacity = '1'
-    setDraggedItem(null)
-    setDragOverItem(null)
-    setDragSection(null)
-    dragCounter.current = 0
-  }
-
-  const handleDragEnter = (e: React.DragEvent, itemId: string) => {
-    e.preventDefault()
-    dragCounter.current++
-    if (itemId !== draggedItem) {
-      setDragOverItem(itemId)
-    }
-  }
-
-  const handleDragLeave = () => {
-    dragCounter.current--
-    if (dragCounter.current === 0) {
-      setDragOverItem(null)
-    }
-  }
-
-  const handleDragOver = (e: React.DragEvent) => {
-    e.preventDefault()
-    e.dataTransfer.dropEffect = 'move'
-  }
-
-  const handleDrop = (e: React.DragEvent, targetId: string, sectionId: string) => {
-    e.preventDefault()
-    dragCounter.current = 0
-
-    if (!draggedItem || draggedItem === targetId || sectionId !== dragSection) {
-      setDraggedItem(null)
-      setDragOverItem(null)
-      setDragSection(null)
-      return
-    }
-
-    // Map section IDs back to the sidebar config sections
-    const section = sectionId as 'primary' | 'secondary'
-    const items = section === 'primary' ? [...config.primaryNav] : [...config.secondaryNav]
-    const draggedIndex = items.findIndex(item => item.id === draggedItem)
-    const targetIndex = items.findIndex(item => item.id === targetId)
-
-    if (draggedIndex === -1 || targetIndex === -1) return
-
-    const [removed] = items.splice(draggedIndex, 1)
-    items.splice(targetIndex, 0, removed)
-
-    const reorderedItems = items.map((item, index) => ({ ...item, order: index }))
-    reorderItems(reorderedItems, section)
-
-    setDraggedItem(null)
-    setDragOverItem(null)
-    setDragSection(null)
-  }
-
-  // ---- Rendering helpers ----
-  const renderIcon = (iconName: string, className?: string) => {
-    const IconComponent = iconRegistry[iconName] as React.ComponentType<{ className?: string }> | undefined
-    return IconComponent ? <IconComponent className={className} /> : null
-  }
-
-  const getCompactBadgeLabel = (badgeValue: string) => {
-    const numericBadge = Number(badgeValue)
-
-    if (!Number.isFinite(numericBadge)) {
-      return badgeValue
-    }
-
-    return numericBadge > COLLAPSED_BADGE_MAX_COUNT
-      ? `${COLLAPSED_BADGE_MAX_COUNT}+`
-      : String(numericBadge)
-  }
-
-  const getNavItemBadge = (item: SidebarNavItem) => {
-    const dashboardId = HREF_TO_DASHBOARD_ID[item.href]
-    const count = dashboardId
-      ? getSidebarCardCount(DASHBOARD_CONFIGS[dashboardId])
-      : null
-    const rawBadge = item.badge ?? (count != null ? String(count) : null)
-    const trimmedBadge = rawBadge?.trim()
-
-    if (!trimmedBadge) {
-      return {
-        count,
-        compactLabel: null,
-        tooltipLabel: null,
-        colorClassName: item.badgeColor ?? COLLAPSED_BADGE_DEFAULT_COLOR_CLASS,
-      }
-    }
-
-    const numericBadge = Number(trimmedBadge)
-    const hasNumericBadge = Number.isFinite(numericBadge)
-
-    if (hasNumericBadge && numericBadge <= 0) {
-      return {
-        count,
-        compactLabel: null,
-        tooltipLabel: null,
-        colorClassName: item.badgeColor ?? COLLAPSED_BADGE_DEFAULT_COLOR_CLASS,
-      }
-    }
-
-    return {
-      count,
-      compactLabel: getCompactBadgeLabel(trimmedBadge),
-      tooltipLabel: trimmedBadge,
-      colorClassName: item.badgeColor ?? COLLAPSED_BADGE_DEFAULT_COLOR_CLASS,
-    }
-  }
-
-  /* Disable drag-reorder on mobile — draggable elements intercept touch
-   * events on Safari, preventing NavLink taps from registering. */
+  const handleCancelRename = () => { setEditingItemId(null); setEditingName('') }
+  const handleClusterStatusClick = (status: 'healthy' | 'unhealthy' | 'unreachable') => navigate(`${ROUTES.CLUSTERS}?status=${status}`)
   const canDrag = features.dragReorder !== false && !isMobile
-
-  const renderNavItem = (item: SidebarNavItem, sectionId: string) => {
-    const isEditing = editingItemId === item.id
-    const navItemBadge = getNavItemBadge(item)
-    const showTooltip = isCollapsed && !isEditing
-    const tooltipBadgeText = navItemBadge.tooltipLabel
-      ? ` (${navItemBadge.tooltipLabel})`
-      : ''
-    const tooltipContent = showTooltip
-      ? `${item.label}${tooltipBadgeText} — ${t('help.sidebarNavItem')}`
-      : ''
-    const navItemTitle = [
-      navItemBadge.tooltipLabel
-        ? `${item.label} (${navItemBadge.tooltipLabel})`
-        : item.label,
-    ]
-
-    if (item.isCustom && item.href.startsWith('/custom-dashboard/')) {
-      navItemTitle.push(t('sidebar.doubleClickRename'))
-    }
-
-    return (
-      <Tooltip
-        key={item.id}
-        content={tooltipContent}
-        side="right"
-        disabled={!showTooltip}
-        wrapperClassName="block w-full"
-      >
-      <div
-        draggable={canDrag && !isCollapsed && !isEditing}
-        onDragStart={(e) => handleDragStart(e, item.id, sectionId)}
-        onDragEnd={handleDragEnd}
-        onDragEnter={(e) => handleDragEnter(e, item.id)}
-        onDragLeave={handleDragLeave}
-        onDragOver={handleDragOver}
-        onDrop={(e) => handleDrop(e, item.id, sectionId)}
-        className={cn(
-          'group relative transition-opacity duration-150 w-full',
-          dragOverItem === item.id && dragSection === sectionId && 'before:absolute before:inset-x-0 before:-top-0.5 before:h-0.5 before:bg-purple-500 before:rounded-full',
-          draggedItem === item.id && 'opacity-50'
-        )}
-      >
-        {isEditing ? (
-          <div className={cn(
-            'flex items-center gap-3 rounded-lg text-sm font-medium',
-            'bg-purple-500/20 text-purple-400',
-            isCollapsed ? 'justify-center p-3' : 'px-3 py-2'
-          )}>
-            {renderIcon(item.icon, isCollapsed ? 'w-6 h-6' : 'w-5 h-5')}
-            {!isCollapsed && (
-              <input
-                type="text"
-                value={editingName}
-                onChange={(e) => setEditingName(e.target.value)}
-                onBlur={() => handleSaveRename(item.id)}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter') handleSaveRename(item.id)
-                  if (e.key === 'Escape') { setEditingItemId(null); setEditingName('') }
-                }}
-                autoFocus
-                className="w-[150px] md:w-full md:flex-1 shrink bg-transparent border-b border-purple-500 outline-hidden text-foreground text-sm min-w-0"
-              />
-            )}
-            {!isCollapsed && <GripVertical className="w-3.5 h-3.5 text-muted-foreground/50 shrink-0" />}
-          </div>
-        ) : (
-          <>
-          <NavLink
-            to={item.href}
-            data-testid="sidebar-item"
-            data-test-label={item.label}
-            onClick={() => emitSidebarNavigated(item.href)}
-            onDoubleClick={(e) => handleDoubleClick(item, e)}
-            onMouseEnter={() => prefetchDashboard(item.href)}
-            className={({ isActive }) => cn(
-              'relative flex items-center gap-3 rounded-lg text-sm font-medium transition-colors duration-200',
-              isActive
-                ? 'bg-purple-500/15 text-purple-400 border-l-[3px] border-purple-500 shadow-purple-inset'
-                : 'text-muted-foreground hover:text-foreground hover:bg-secondary/50 border-l-[3px] border-transparent',
-              isCollapsed ? 'justify-center p-3' : 'px-3 py-2'
-            )}
-            title={(navItemTitle || []).join(' — ')}
-          >
-            <span className="relative shrink-0">
-              {renderIcon(item.icon, isCollapsed ? 'w-6 h-6' : 'w-5 h-5')}
-              {isCollapsed && navItemBadge.compactLabel && (
-                <span
-                  className={cn(COLLAPSED_BADGE_BASE_CLASS, navItemBadge.colorClassName)}
-                  aria-hidden="true"
-                >
-                  {navItemBadge.compactLabel}
-                </span>
-              )}
-            </span>
-            {!isCollapsed && (() => {
-              const isGC = isGroundControlItem(item.href)
-              return (
-                <span className="flex-1 min-w-0 flex items-center gap-1">
-                  <span className="truncate">{item.label}</span>
-                  {isGC && (
-                    <Satellite className="w-3.5 h-3.5 text-purple-400 shrink-0" aria-label="Ground Control dashboard" />
-                  )}
-                  {navItemBadge.count != null && (
-                    <span
-                      className="text-xs text-muted-foreground/40 tabular-nums ml-0.5 shrink-0"
-                      title={t('sidebar.cardCount', { count: navItemBadge.count })}
-                    >{navItemBadge.count}</span>
-                  )}
-                </span>
-              )
-            })()}
-          </NavLink>
-            {!isCollapsed && canDrag && (
-              <span className="absolute right-5 top-1/2 -translate-y-1/2 flex items-center gap-1 opacity-100 md:opacity-0 md:group-hover:opacity-100 transition-opacity bg-background/80 backdrop-blur-xs rounded px-1 z-10">
-                {!PROTECTED_SIDEBAR_IDS.includes(item.id) && (
-                  <button
-                    type="button"
-                    onClick={(e) => { e.preventDefault(); e.stopPropagation(); removeItem(item.id) }}
-                    className="p-1 rounded hover:bg-red-500/20 hover:text-red-400 text-muted-foreground/50 transition-colors"
-                    title={t('sidebar.removeFromSidebar')}
-                    aria-label={t('sidebar.removeFromSidebar')}
-                  >
-                    <X className="w-3.5 h-3.5" aria-hidden="true" />
-                  </button>
-                )}
-                <span
-                  className="p-1 rounded hover:bg-secondary text-muted-foreground/50 hover:text-muted-foreground cursor-grab active:cursor-grabbing transition-colors"
-                  aria-hidden="true"
-                  onMouseDown={(e) => e.stopPropagation()}
-                >
-                  <GripVertical
-                    className="w-4 h-4"
-                  />
-                </span>
-              </span>
-            )}
-          </>
-        )}
-      </div>
-      </Tooltip>
-    )
-  }
 
   /** Render a collapsible nav section with header */
   const [collapsedSections, setCollapsedSections] = useState<Record<string, boolean>>({})
@@ -667,10 +256,8 @@ export function SidebarShell({
 
     return (
       <div key={section.id}>
-        {/* Divider between sections (except before the first) */}
         {index > 0 && <div className="my-6 border-t border-border/50" />}
 
-        {/* Collapsible section header */}
         {section.label && !isCollapsed && (
           <button
             onClick={() => section.collapsible && toggleSection(section.id)}
@@ -682,12 +269,13 @@ export function SidebarShell({
           >
             <span className="flex-1 text-left">{section.label}</span>
             {section.collapsible && (
-              isOpen ? <ChevronDown className="w-3 h-3" /> : <ChevronRight className="w-3 h-3" />
+              isOpen
+                ? <span className="text-xs">▾</span>
+                : <span className="text-xs">▸</span>
             )}
           </button>
         )}
 
-        {/* Section items */}
         {(isOpen || !section.collapsible) && (
           <nav
             data-testid={`sidebar-${section.id}-nav`}
@@ -696,7 +284,31 @@ export function SidebarShell({
               moveFocusByKey(event, { selector: 'a[data-testid="sidebar-item"]', orientation: 'vertical' })
             }}
           >
-            {section.items.map(item => renderNavItem(item, section.id))}
+            {section.items.map(item => (
+              <SidebarNavItemRow
+                key={item.id}
+                item={item}
+                sectionId={section.id}
+                isCollapsed={isCollapsed}
+                canDrag={canDrag}
+                editingItemId={editingItemId}
+                editingName={editingName}
+                draggedItem={draggedItem}
+                dragOverItem={dragOverItem}
+                dragSection={dragSection}
+                onDragStart={handleDragStart}
+                onDragEnd={handleDragEnd}
+                onDragEnter={handleDragEnter}
+                onDragLeave={handleDragLeave}
+                onDragOver={handleDragOver}
+                onDrop={handleDrop}
+                onDoubleClick={handleDoubleClick}
+                onSaveRename={handleSaveRename}
+                onCancelRename={handleCancelRename}
+                onRenameChange={setEditingName}
+                onRemove={removeItem}
+              />
+            ))}
           </nav>
         )}
       </div>
@@ -806,108 +418,26 @@ export function SidebarShell({
 
         {/* Cluster status summary */}
         {features.clusterStatus && !isCollapsed && (
-          <div data-testid="sidebar-cluster-status" className="mt-6 p-4 rounded-lg bg-secondary/30">
-            <h4 className="text-xs font-medium text-muted-foreground uppercase tracking-wider mb-3">
-              {t('labels.clusterStatus')}
-            </h4>
-            {deduplicatedClusters.length === 0 ? (
-              <p className="text-xs text-muted-foreground">{t('labels.noClusters')}</p>
-            ) : (
-            <div className="space-y-2">
-              {healthyClusters > 0 && (
-              <button
-                onClick={() => handleClusterStatusClick('healthy')}
-                className="w-full flex items-center justify-between hover:bg-secondary/50 rounded px-1 py-0.5 transition-colors"
-              >
-                <span className="flex items-center gap-1.5 text-sm text-foreground">
-                  <CheckCircle2 className="w-3.5 h-3.5 text-green-400" aria-hidden="true" />
-                  {t('labels.healthy')}
-                </span>
-                <span
-                  className="text-sm font-medium text-green-400"
-                  title={t('sidebar.healthyClusters', { count: healthyClusters })}
-                >{healthyClusters}</span>
-              </button>
-              )}
-              {unhealthyClusters > 0 && (
-              <button
-                onClick={() => handleClusterStatusClick('unhealthy')}
-                className="w-full flex items-center justify-between hover:bg-secondary/50 rounded px-1 py-0.5 transition-colors"
-              >
-                <span className="flex items-center gap-1.5 text-sm text-foreground">
-                  <AlertTriangle className="w-3.5 h-3.5 text-red-400" aria-hidden="true" />
-                  {t('labels.unhealthy')}
-                </span>
-                <span
-                  className="text-sm font-medium text-red-400"
-                  title={t('sidebar.unhealthyClusters', { count: unhealthyClusters })}
-                >{unhealthyClusters}</span>
-              </button>
-              )}
-              {unreachableClusters > 0 && (
-              <button
-                onClick={() => handleClusterStatusClick('unreachable')}
-                className="w-full flex items-center justify-between hover:bg-secondary/50 rounded px-1 py-0.5 transition-colors"
-              >
-                <span className="flex items-center gap-1.5 text-sm text-foreground">
-                  <WifiOff className="w-3.5 h-3.5 text-yellow-400" aria-hidden="true" />
-                  {t('labels.offline')}
-                </span>
-                <span
-                  className="text-sm font-medium text-yellow-400"
-                  title={t('sidebar.unreachableClusters', { count: unreachableClusters })}
-                >{unreachableClusters}</span>
-              </button>
-              )}
-              {healthyClusters === 0 && unhealthyClusters === 0 && unreachableClusters === 0 && (
-                <span className="text-xs text-muted-foreground italic">{t('labels.noClusters', 'No clusters configured')}</span>
-              )}
-            </div>
-            )}
-          </div>
+          <SidebarClusterStatus
+            healthyClusters={healthyClusters}
+            unhealthyClusters={unhealthyClusters}
+            unreachableClusters={unreachableClusters}
+            onStatusClick={handleClusterStatusClick}
+          />
         )}
 
-        {/* Viewer count + commit hash — separated from cluster status to prevent
-          * the commit SHA from visually merging with cluster counts (#11403). */}
+        {/* Viewer count + commit hash */}
         {features.activeUsers && !isCollapsed && (
-          <div className="mt-auto pt-4 border-t border-border/30 flex flex-col items-center gap-1">
-            <div className="flex items-center justify-center gap-2">
-              <div className="flex items-center gap-1 px-2 text-muted-foreground/60">
-                <span className="sr-only">{t('sidebar.activeViewers', { count: viewerCount })}</span>
-                <User className={cn('w-3 h-3', viewersError && 'text-red-400')} aria-hidden="true" />
-                <span className="text-2xs tabular-nums" aria-hidden="true">
-                  {viewersError ? '!' : viewersLoading ? '…' : viewerCount}
-                </span>
-              </div>
-              <span className="text-2xs text-muted-foreground/40 font-mono" title={`Commit: ${__COMMIT_HASH__}`}>
-                <span className="sr-only">{`Commit: ${__COMMIT_HASH__}`}</span>
-                <span aria-hidden="true">#{__COMMIT_HASH__.substring(0, 7)}</span>
-              </span>
-            </div>
-            {/* Developer mode: warn when running an older commit, or show upgrade progress */}
-            {features.versionCheck && channel === 'developer' && hasUpdate && (
-              <div
-                className={cn(
-                  'flex items-center gap-1 text-2xs',
-                  isUpgrading ? 'text-cyan-400/80' : 'text-yellow-400/80',
-                )}
-                title={isUpgrading
-                  ? t('update.upgrading', 'Upgrading...')
-                  : `Behind main — latest: ${latestMainSHA?.substring(0, 7) ?? 'unknown'}`}
-              >
-                {isUpgrading ? (
-                  <Loader2 className="w-3 h-3 animate-spin" />
-                ) : (
-                  <AlertTriangle className="w-3 h-3" />
-                )}
-                <span>
-                  {isUpgrading
-                    ? t('update.upgrading', 'Upgrading...')
-                    : t('sidebar.behindMain', 'Behind main')}
-                </span>
-              </div>
-            )}
-          </div>
+          <SidebarActiveUsersFooter
+            viewerCount={viewerCount}
+            viewersError={!!viewersError}
+            viewersLoading={viewersLoading}
+            showVersionCheck={features.versionCheck ?? false}
+            channel={channel}
+            hasUpdate={hasUpdate}
+            isUpgrading={isUpgrading}
+            latestMainSHA={latestMainSHA}
+          />
         )}
 
         {/* Custom footer */}
@@ -915,47 +445,21 @@ export function SidebarShell({
       </aside>
 
       {/* Collapse + Pin controls */}
-      {features.collapsePin !== false && !isMobile && !isMissionFullScreen && (
-        <div
-          className="fixed top-18 z-sidebar flex flex-col gap-1.5 items-center transition-[left] duration-300 bg-background border border-border/50 rounded-full p-1 shadow-md"
-          style={{ left: sidebarWidth + SIDEBAR_CONTROLS_LEFT_OFFSET_PX }}
-        >
-          <button
-            data-testid="sidebar-collapse-toggle"
-            onClick={() => {
-              if (config.collapsed) {
-                setCollapsed(false)
-                if (!isPinned) {
-                  setIsPinned(true)
-                  try { localStorage.setItem('sidebar-left-pinned', 'true') } catch { /* ignore */ }
-                  clearAutoHideTimer()
-                }
-              } else {
-                toggleCollapsed()
-              }
-            }}
-            aria-expanded={!config.collapsed}
-            className="hidden md:flex items-center justify-center w-8 h-8 rounded-full bg-secondary text-muted-foreground hover:text-foreground hover:bg-secondary/80 transition-colors"
-            title={config.collapsed ? t('layout.sidebar.expandSidebar') : t('layout.sidebar.collapseSidebar')}
-            aria-label={config.collapsed ? t('layout.sidebar.expandSidebar') : t('layout.sidebar.collapseSidebar')}
-          >
-            {config.collapsed ? <ChevronRight className="w-3.5 h-3.5" aria-hidden="true" /> : <ChevronLeft className="w-3.5 h-3.5" aria-hidden="true" />}
-          </button>
-          <button
-            onClick={toggleSidebarPin}
-            aria-pressed={isPinned}
-            className={cn(
-              "flex items-center justify-center w-8 h-8 rounded-full transition-colors",
-              isPinned
-                ? "bg-purple-500/15 text-purple-400 hover:bg-purple-500/25"
-                : "bg-secondary text-muted-foreground hover:text-foreground hover:bg-secondary/80"
-            )}
-            title={isPinned ? t('layout.sidebar.unpinSidebar') : t('layout.sidebar.pinSidebar')}
-            aria-label={isPinned ? t('layout.sidebar.unpinSidebar') : t('layout.sidebar.pinSidebar')}
-          >
-            {isPinned ? <Pin className="w-3.5 h-3.5" aria-hidden="true" /> : <PinOff className="w-3.5 h-3.5" aria-hidden="true" />}
-          </button>
-        </div>
+      {features.collapsePin !== false && !isMobile && (
+        <SidebarCollapseControls
+          isCollapsed={isCollapsed}
+          isPinned={isPinned}
+          sidebarWidth={sidebarWidth}
+          isMissionFullScreen={isMissionFullScreen}
+          onToggleCollapse={() => {
+            if (config.collapsed) {
+              setCollapsed(false)
+            } else {
+              toggleCollapsed()
+            }
+          }}
+          onTogglePin={toggleSidebarPin}
+        />
       )}
 
       {/* Resize handle */}
