+++ web/src/components/layout/sidebar/useSidebarDragDrop.ts
@@ -1,0 +1,101 @@
+import { useState, useRef } from 'react'
+
+interface NavItem {
+  id: string
+  order: number
+  [key: string]: unknown
+}
+
+interface UseSidebarDragDropParams {
+  primaryNav: NavItem[]
+  secondaryNav: NavItem[]
+  reorderItems: (items: NavItem[], section: 'primary' | 'secondary') => void
+}
+
+export function useSidebarDragDrop({ primaryNav, secondaryNav, reorderItems }: UseSidebarDragDropParams) {
+  const [draggedItem, setDraggedItem] = useState<string | null>(null)
+  const [dragOverItem, setDragOverItem] = useState<string | null>(null)
+  const [dragSection, setDragSection] = useState<string | null>(null)
+  const dragCounter = useRef(0)
+
+  const handleDragStart = (e: React.DragEvent, itemId: string, sectionId: string) => {
+    setDraggedItem(itemId)
+    setDragSection(sectionId)
+    e.dataTransfer.effectAllowed = 'move'
+    e.dataTransfer.setData('text/plain', itemId)
+    requestAnimationFrame(() => {
+      const target = e.target as HTMLElement
+      target.style.opacity = '0.5'
+    })
+  }
+
+  const handleDragEnd = (e: React.DragEvent) => {
+    const target = e.target as HTMLElement
+    target.style.opacity = '1'
+    setDraggedItem(null)
+    setDragOverItem(null)
+    setDragSection(null)
+    dragCounter.current = 0
+  }
+
+  const handleDragEnter = (e: React.DragEvent, itemId: string) => {
+    e.preventDefault()
+    dragCounter.current++
+    if (itemId !== draggedItem) {
+      setDragOverItem(itemId)
+    }
+  }
+
+  const handleDragLeave = () => {
+    dragCounter.current--
+    if (dragCounter.current === 0) {
+      setDragOverItem(null)
+    }
+  }
+
+  const handleDragOver = (e: React.DragEvent) => {
+    e.preventDefault()
+    e.dataTransfer.dropEffect = 'move'
+  }
+
+  const handleDrop = (e: React.DragEvent, targetId: string, sectionId: string) => {
+    e.preventDefault()
+    dragCounter.current = 0
+
+    if (!draggedItem || draggedItem === targetId || sectionId !== dragSection) {
+      setDraggedItem(null)
+      setDragOverItem(null)
+      setDragSection(null)
+      return
+    }
+
+    const section = sectionId as 'primary' | 'secondary'
+    const items = section === 'primary' ? [...primaryNav] : [...secondaryNav]
+    const draggedIndex = items.findIndex(item => item.id === draggedItem)
+    const targetIndex = items.findIndex(item => item.id === targetId)
+
+    if (draggedIndex === -1 || targetIndex === -1) return
+
+    const [removed] = items.splice(draggedIndex, 1)
+    items.splice(targetIndex, 0, removed)
+
+    const reorderedItems = items.map((item, index) => ({ ...item, order: index }))
+    reorderItems(reorderedItems, section)
+
+    setDraggedItem(null)
+    setDragOverItem(null)
+    setDragSection(null)
+  }
+
+  return {
+    draggedItem,
+    dragOverItem,
+    dragSection,
+    handleDragStart,
+    handleDragEnd,
+    handleDragEnter,
+    handleDragLeave,
+    handleDragOver,
+    handleDrop,
+  }
+}
